Completely restructure note track scanning
Instead of holding each separate scan with it's type denoted explicitly, only make it hold the subtrack and intensity values. Use free template functions specialized for each noteTrack type and separated based on the chart being parsed to extract the valid values.

Essentially, make it so the libraryEntry structure knows the type of note each value lines up with, but *not* the structure of each note. Encapsulation.